requirements.txt
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pip-review==1.3.0
scipy==1.14.1
